+++ src/ChatWootApi/Application/Models/ReportingEventMeta.cs
@@ -12,19 +12,19 @@ public sealed record ReportingEventMeta
     /// 报告事件总数
     /// </summary>
     [JsonPropertyName("count")]
-    public long? Count { get; init; }
+    public int? Count { get; init; }
 
     /// <summary>
     /// 当前页码
     /// </summary>
     [JsonPropertyName("current_page")]
-    public long? CurrentPage { get; init; }
+    public int? CurrentPage { get; init; }
 
     /// <summary>
     /// 总页数
     /// </summary>
     [JsonPropertyName("total_pages")]
-    public long? TotalPages { get; init; }
+    public int? TotalPages { get; init; }
 
     /// <summary>
     /// Swagger 未显式建模的附加 JSON 字段
